+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
-- Use Heartbeat v2 API
+### Fixed
+- Use Heartbeat v2 API, as the Heartbeat v1 API is no longer available
 
 ## [3.1.0] - 2017-11-12
 ### Added